requirements.txt
mysql-connector-python==8.0.29
pyTelegramBotAPI==4.16.1
requests==2.28.0
idna==3.3
urllib3==1.26.9